refactor: delete legacy service files replaced by RTK Query slices
- Remove 9 legacy axios-based service files: settingsApi, printSettingsApi,
  paymentMethodsApi, inventoryApi, salesApi, purchasingApi, backupService,
  auditLogApi, priceListApi
- Move BulkUpdatePriceDto, CopyPriceListDto, PercentageAdjustmentDto type
  definitions into store/api/priceListApi.ts (RTK slice)
- Migrate AuditLogsPage from legacy priceListApi to RTK slice initiate()
- Migrate PriceCostingPage from legacy inventoryApi to direct ApiService call
- Migrate BackupList from backupService.downloadBackup to ApiService.downloadFile
